Working animation (except for Random button...)
So, instead of using an ObjectAnimator which modifies the property of its target, we use a ValueAnimator, from which we get the value and we set that value on the target ourselves. The target is manually updated when we bind/unbind a ViewHolder to a Sound. There are some leftover code from previous attempts; needs cleanup